refactor(logistica): validação completa do módulo com rastreamento de entregas e integração com pedidos
