+++ AGENTS.md
@@ -75,3 +75,4 @@ Data | Autor | Descrição
 2025-06-10 | CODEX | Módulo de produção validado com Kanban e rastreabilidade.
 2025-06-10 | CODEX | Módulo de compras validado com integração estoque/financeiro.
 2025-06-10 | CODEX | Módulo financeiro criado com integração de receitas e despesas.
+2025-06-10 | CODEX | Módulo de logística validado com rastreamento de entregas.
