+++ CHECKLIST.md
@@ -70,3 +70,4 @@ Este arquivo centraliza os checklists de tarefas, validações e revisões para 
 - 2025-06-10: Módulo de produção validado com Kanban e rastreabilidade (CODEX)
 - 2025-06-10: Módulo de compras revisado e validado (CODEX)
 - 2025-06-10: Módulo financeiro criado e integrado (CODEX)
+- 2025-06-10: Módulo de logística validado com rastreamento de entregas (CODEX)
