+++ src/modules/logistica/EntregaForm.tsx
@@ -1,0 +1,1 @@
+export { DeliveryForm as EntregaForm } from '@/app/(dashboard)/logistica/_components/DeliveryForm';
